claudebot/requirements.txt
anthropic>=0.40.0
python-telegram-bot>=20.0
